requirements.txt
fastapi 
uvicorn 
apscheduler 
httpx   
sqlalchemy 
aiosqlite 
python-dotenv 
pydantic   
httpx 
requests 
jinja2
